refactor: fetch 함수 프록시 처리
