+++ src/app/media/highlight/[videoId]/page.tsx
@@ -2,7 +2,7 @@ import BookmarkButton from '@/components/media/video/bookmark-button'
 import RelatedVideoList from '@/components/media/video/related-video-list'
 import VideoEmbed from '@/components/media/video/video-embed'
 import Breadcrumbs from '@/components/tailwind-ui/breadcrumbs/simple-with-chevrons'
-import { checkBookmarked } from '@/services/media-action'
+import { checkBookmarked } from '@/services/bookmark-action'
 import { getVideoDetail } from '@/services/media-service'
 import { Video } from '@/types/media'
 import { CalendarDaysIcon, CheckCircleIcon } from '@heroicons/react/24/outline'
